+++ src/main/java/fastcampus/boardproject/SpringbootBoardProjectApplication.java
@@ -1,4 +1,4 @@
-package fastcapus.boardproject;
+package fastcampus.boardproject;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
