+++ src/main/java/fastcampus/boardproject/domain/Article.java
@@ -1,11 +1,12 @@
-package fastcapus.boardproject.domain;
+package fastcampus.boardproject.domain;
 
 import java.time.LocalDateTime;
 
-public class ArticleComment {
+public class Article {
     private Long id;
-    private Article article; // 게시글 (ID)
+    private String title; // 제목
     private String content; // 본문
+    private String hashtag; //해시태그
 
     private LocalDateTime createdAt; // 생성일시
     private String createdBy; // 생성자
