+++ src/main/java/fastcampus/boardproject/domain/ArticleComment.java
@@ -1,12 +1,11 @@
-package fastcapus.boardproject.domain;
+package fastcampus.boardproject.domain;
 
 import java.time.LocalDateTime;
 
-public class Article {
+public class ArticleComment {
     private Long id;
-    private String title; // 제목
+    private Article article; // 게시글 (ID)
     private String content; // 본문
-    private String hashtag; //해시태그
 
     private LocalDateTime createdAt; // 생성일시
     private String createdBy; // 생성자
