+++ src/main/resources/application.yaml
@@ -1,0 +1,24 @@
+debug: false
+management.endpoint.web.exposure.include: "*"
+
+logging:
+  level:
+    com.fastcampus.boardproject: debug
+    org.springframework.web.servlet: debug
+    org.hibernate.type.descriptor.sql.BasicBinder: trace
+
+spring:
+  datasource:
+    url: jdbc:mysql//localhost:3306/board
+    username: orchid
+    password: P@ssw0rd1!
+    driver-class-name: com.mysql.cj.jdbc.Driver
+  jpa:
+    defer-datasource-initialization: true
+    hibernate.ddl-auto: create
+    open-in-view: false
+    show-sql: true
+    properties:
+      hibernate.format_sql: true
+      hibernate.default_fetch_size: 100
+    h2.console.enabled: false
